requirements.txt
alembic==1.4.2
Babel==2.8.0
blinker==1.4
click==7.1.1
enum34==1.1.10
Flask==1.1.2
Flask-Admin==1.5.6
Flask-BabelEx==0.9.4
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.1
Flask-SSLify==0.1.5
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
passlib==1.7.2
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
six==1.14.0
speaklater==1.3
SQLAlchemy==1.3.16
typing==3.7.4.1
Werkzeug==1.0.1
WTForms==2.2.1
